fixed for delete without attachment

+++ app/Http/Controllers/TransactionController.php
@@ -254,7 +254,10 @@ class TransactionController extends Controller
     public function destroy($id)
     {
         $transaction = Transaction::where('id',$id);
-        Storage::delete($transaction->attachment);
+        if($transaction->attachment !=null && $transaction->attachment != "") {
+            Storage::delete($transaction->attachment);
+        }
+        
         $transaction->delete();
         return redirect()->route('transactions.index');
     }
